Fuse fmpz_mpoly/evaluate*
